[Nanchen/Shenjian][#136] fixed compressing image error when width < height.

+++ app/src/main/java/org/unicef/rapidreg/utils/ImageCompressUtil.java
@@ -76,7 +76,7 @@ public class ImageCompressUtil {
         int imageHeight = options.outHeight;
         int imageWidth = options.outWidth;
 
-        float imgRatio = imageWidth / imageHeight;
+        float imgRatio = imageWidth / (float) imageHeight;
         float maxRatio = maxWidth / maxHeight;
 
         if (imageHeight > maxHeight || imageWidth > maxWidth) {
